backend/requirements.txt
flask
flask-cors
psutil
watchdog
scikit-learn
pandas
numpy
joblib
supabase
python-dotenv
requests
validators
python-magic-bin
